+++ Backend/database/migrations/2025_01_16_095826_add_idea_path_id_column_to_mentors_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('mentors', function (Blueprint $table) {
+        if (Schema::hasTable('mentors')) {
+            Schema::table('mentors', function (Blueprint $table) {
             $table->foreignId('track_id')->constrained('paths')->cascadeOnDelete();
         });
+        }
     }
 
     /**
@@ -21,9 +23,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('mentors', function (Blueprint $table) {
+        if (Schema::hasTable('mentors')) {
+            Schema::table('mentors', function (Blueprint $table) {
             $table->dropForeign(['track_id']);
             $table->dropColumn('track_id');
         });
+        }
     }
 };
