+++ Backend/database/migrations/2025_03_12_143323_add_section_id_to_form_fields_table.php
@@ -10,13 +10,15 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::table('form_fields', function (Blueprint $table) {
+        if (Schema::hasTable('form_fields')) {
+            Schema::table('form_fields', function (Blueprint $table) {
             $table->foreignId('section_id')
                 ->nullable()
                 
                 ->constrained('form_sections')
                 ->onDelete('cascade');
         });
+        }
     }
 
     /**
@@ -24,8 +26,10 @@ return new class extends Migration {
      */
     public function down(): void
     {
-        Schema::table('form_fields', function (Blueprint $table) {
+        if (Schema::hasTable('form_fields')) {
+            Schema::table('form_fields', function (Blueprint $table) {
             //
         });
+        }
     }
 };
